Separate lightning utils.
