+++ mart/utils/lightning.py
